+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.2.0
 
 - pkger doesn't start a build by default, there is a separate subcommand `build` for that now. [#22](https://github.com/wojciechkepka/pkger/pull/22)
+- Add `gen-recipe` subcommand to declaratively generate recipes [#22](https://github.com/wojciechkepka/pkger/pull/22)